+++ src/main/java/com/besson/arknights/block/custom/lgd/BlackVintageStool.java
@@ -1,0 +1,44 @@
+package com.besson.arknights.block.custom.lgd;
+
+import com.besson.arknights.block.FurnitureHorizontalFacingBlock;
+import com.besson.arknights.block.SeatEntity;
+import net.minecraft.block.Block;
+import net.minecraft.block.BlockState;
+import net.minecraft.block.ShapeContext;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.util.ActionResult;
+import net.minecraft.util.Hand;
+import net.minecraft.util.function.BooleanBiFunction;
+import net.minecraft.util.hit.BlockHitResult;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.shape.VoxelShape;
+import net.minecraft.util.shape.VoxelShapes;
+import net.minecraft.world.BlockView;
+import net.minecraft.world.World;
+
+import java.util.stream.Stream;
+
+public class BlackVintageStool extends FurnitureHorizontalFacingBlock {
+    public BlackVintageStool(Settings settings) {
+        super(settings);
+    }
+    private static final VoxelShape SHAPE = Stream.of(
+            Block.createCuboidShape(2, 0, 2, 4, 8, 4),
+            Block.createCuboidShape(12, 0, 2, 14, 8, 4),
+            Block.createCuboidShape(12, 0, 12, 14, 8, 14),
+            Block.createCuboidShape(2, 0, 12, 4, 8, 14),
+            Block.createCuboidShape(2, 8, 2, 14, 10, 14)
+    ).reduce((v1, v2) -> VoxelShapes.combineAndSimplify(v1, v2, BooleanBiFunction.OR)).get();
+
+    @Override
+    public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
+        return SHAPE;
+    }
+    @Override
+    public ActionResult onUse(BlockState state, World world, BlockPos pos, PlayerEntity player, Hand hand, BlockHitResult hit) {
+        if (!world.isClient){
+            return SeatEntity.create(world,pos,0.4,player,state.get(FACING));
+        }
+        return ActionResult.SUCCESS;
+    }
+}