+++ src/main/java/com/besson/arknights/block/custom/lgd/SmallSandalwoodNightstand.java
@@ -1,0 +1,33 @@
+package com.besson.arknights.block.custom.lgd;
+
+import com.besson.arknights.block.FurnitureHorizontalFacingBlock;
+import com.besson.arknights.block.ModAbstractChestBlock;
+import net.minecraft.block.Block;
+import net.minecraft.block.BlockState;
+import net.minecraft.block.ShapeContext;
+import net.minecraft.block.entity.BlockEntity;
+import net.minecraft.block.entity.BlockEntityType;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.shape.VoxelShape;
+import net.minecraft.world.BlockView;
+import org.jetbrains.annotations.Nullable;
+
+import java.util.function.Supplier;
+
+public class SmallSandalwoodNightstand extends ModAbstractChestBlock<SmallSandalwoodNightstandBE> {
+
+    public SmallSandalwoodNightstand(Settings settings, Supplier<BlockEntityType<? extends SmallSandalwoodNightstandBE>> blockEntityTypeSupplier) {
+        super(settings, blockEntityTypeSupplier);
+    }
+    private static final VoxelShape SHAPE = Block.createCuboidShape(0, 0, 0, 16, 16, 16);
+    @Override
+    public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
+        return SHAPE;
+    }
+
+    @Nullable
+    @Override
+    public BlockEntity createBlockEntity(BlockPos pos, BlockState state) {
+        return new SmallSandalwoodNightstandBE(pos, state);
+    }
+}